reorg ./pkg/logs/parser into more files
