+++ pkg/logs/parser/decoding.go
@@ -1,0 +1,49 @@
+package parser
+
+import (
+	"golang.org/x/text/encoding"
+	"golang.org/x/text/encoding/unicode"
+	"golang.org/x/text/transform"
+)
+
+// Encoding specifies the encoding which should be decoded by the DecodingParser
+type Encoding int
+
+const (
+	// UTF16LE UTF16 little endian, most common (windows)
+	UTF16LE Encoding = iota
+	// UTF16BE UTF16 big endian
+	UTF16BE
+)
+
+// DecodingParser a parser for decoding encoded logfiles.  It treats each input
+// message as entirely content, in the given encoding.  No timetamp or other
+// metadata are returned.
+type DecodingParser struct {
+	decoder *encoding.Decoder
+}
+
+// Parse implements Parser#Parse
+func (p *DecodingParser) Parse(msg []byte) ([]byte, string, string, bool, error) {
+	decoded, _, err := transform.Bytes(p.decoder, msg)
+	return decoded, "", "", false, err
+}
+
+// SupportsPartialLine implements Parser#SupportsPartialLine
+func (p *DecodingParser) SupportsPartialLine() bool {
+	return false
+}
+
+// NewDecodingParser builds a new DecodingParser.
+func NewDecodingParser(e Encoding) *DecodingParser {
+	p := &DecodingParser{}
+	var enc encoding.Encoding
+	switch e {
+	case UTF16LE:
+		enc = unicode.UTF16(unicode.LittleEndian, unicode.UseBOM)
+	case UTF16BE:
+		enc = unicode.UTF16(unicode.BigEndian, unicode.UseBOM)
+	}
+	p.decoder = enc.NewDecoder()
+	return p
+}
